pyproject.toml
[project]
name = "transcriber"
version = "0.2.0"
description = "GPU faster-whisper transcription with direct pyannote diarization (local + Docker)"
authors = [{ name = "rputnam" }]
readme = "README.md"
requires-python = ">=3.10,<3.13"

dependencies = [
    # Runtime ASR + diarization
    "faster-whisper",
    "nemo-toolkit[asr]>=2,<3; sys_platform == 'linux'",
    "pyannote.audio>=4,<5",
    "torch>=2.7,<3",
    "torchaudio>=2.7,<3",
    # Linux x86_64 gets the GPU runtime by default; other platforms stay CPU-safe.
    "onnxruntime-gpu>=1.18,<2; sys_platform == 'linux' and platform_machine == 'x86_64'",
    "onnxruntime>=1.18,<2; sys_platform != 'linux' or platform_machine != 'x86_64'",
    "nvidia-cublas-cu12>=12,<13; sys_platform == 'linux' and platform_machine == 'x86_64'",
    "nvidia-cudnn-cu12>=9,<10; sys_platform == 'linux' and platform_machine == 'x86_64'",
    # Utilities
    "pillow>=10.0.0",
    "pyyaml>=6.0",
    "tqdm>=4.66.0",
    "numpy>=1.24,<3",
    "scikit-learn>=1.4,<2",
    "matplotlib>=3.8,<4",
    # Hugging Face fast transfer (since HF_HUB_ENABLE_HF_TRANSFER=1 is set)
    "hf-transfer>=0.1.8",
    "parakeet-mlx>=0.5.1; sys_platform == 'darwin' and platform_machine == 'arm64'",
]

[project.scripts]
transcribe = "transcriber.cli:main"

[tool.ruff]
line-length = 100

[tool.black]
line-length = 100

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "black>=25,<26",
    "pre-commit>=4,<5",
    "pytest>=8.4,<9",
    "ruff>=0.12,<0.13",
]

[tool.pytest.ini_options]
markers = [
    "integration: opt-in tests that exercise real backend integrations",
    "slow: tests that are intentionally slower than the default suite",
]

# uv: on Linux x86_64, install PyTorch wheels built for CUDA 12.8.
[tool.uv.sources]
torch = [{ index = "pytorch-cu128", marker = "sys_platform == 'linux' and platform_machine == 'x86_64'" }]
torchaudio = [{ index = "pytorch-cu128", marker = "sys_platform == 'linux' and platform_machine == 'x86_64'" }]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
